# api/openapi_stub.yaml
components:
  schemas:
...
    AsyncReportRunResponse:
      properties:
        connector_ingest_run_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Connector Ingest Run Id
        connector_review_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Connector Review Status
        report_run_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Report Run Id
        retry_of_report_run_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Retry Of Report Run Id
        status:
          title: Status
          type: string
      required:
      - status
      title: AsyncReportRunResponse
      type: object
...
    ClaimContract:
      properties:
        area_id:
          format: uuid
          title: Area Id
          type: string
        assertion:
          title: Assertion
          type: string
        claim_code:
          title: Claim Code
          type: string
        claim_id:
          format: uuid
          title: Claim Id
          type: string
        confidence:
          $ref: '#/components/schemas/ConfidenceBand'
          default: unknown
        domain:
          default: unknown
          title: Domain
          type: string
        evidence_ids:
          items:
            format: uuid
            type: string
          title: Evidence Ids
          type: array
        rule_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Rule Code
        ruleset_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Ruleset Id
        ruleset_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Ruleset Version
        severity:
          $ref: '#/components/schemas/SeverityBand'
          default: unknown
        user_safe_language:
          default: ''
          title: User Safe Language
          type: string
        verification_required:
          default: true
          title: Verification Required
          type: boolean
        verification_task:
          anyOf:
          - type: string
          - type: 'null'
          title: Verification Task
      required:
      - area_id
      - claim_code
      - assertion
      - evidence_ids
      title: ClaimContract
      type: object
    ConfidenceBand:
      enum:
      - very_low
      - low
      - medium
      - high
      - very_high
      - unknown
      title: ConfidenceBand
      type: string
...
    EvidenceContract:
      properties:
        area_id:
          format: uuid
          title: Area Id
          type: string
        caveat:
          anyOf:
          - type: string
          - type: 'null'
          title: Caveat
        confidence:
          $ref: '#/components/schemas/ConfidenceBand'
          default: unknown
        dataset_version_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Dataset Version Id
        domain:
          title: Domain
          type: string
        evidence_code:
          title: Evidence Code
          type: string
        evidence_id:
          format: uuid
          title: Evidence Id
          type: string
        evidence_type:
          $ref: '#/components/schemas/EvidenceType'
          default: source_observation
        geometry_geojson:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Geometry Geojson
        geometry_srid:
          default: 4326
          title: Geometry Srid
          type: integer
        is_negative_evidence:
          default: false
          title: Is Negative Evidence
          type: boolean
        is_source_failure:
          default: false
          title: Is Source Failure
          type: boolean
        method_code:
          title: Method Code
          type: string
        method_version:
          default: 0.1.0
          title: Method Version
          type: string
        observation:
          title: Observation
          type: string
        observed_at:
          format: date-time
          title: Observed At
          type: string
        observed_value:
          additionalProperties: true
          title: Observed Value
          type: object
        source_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Date
        source_id:
          format: uuid
          title: Source Id
          type: string
        source_ingest_run_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Source Ingest Run Id
        spatial_precision_meters:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Spatial Precision Meters
        superseded_by:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Superseded By
      required:
      - area_id
      - evidence_code
      - domain
      - observation
      - source_id
      - method_code
      title: EvidenceContract
      type: object
    EvidenceType:
      enum:
      - source_observation
      - spatial_intersection
      - derived_metric
      - document_extract
      - source_failure
      - human_verification
      - manual_note
      title: EvidenceType
      type: string
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    IntentCode:
      description: 'Intent codes that drive report evaluation.


        Values must exactly match the core.intent_code SQL enum defined in

        db/migrations/0001_initial_spine.sql. Two intents are seeded in

        db/seeds/001_seed_intents.sql; others are defined for future use.


        Do not add or rename values here without a coordinated schema migration.'
      enum:
      - rural_land_purchase
      - homestead_feasibility
      - farmland
      - development
      - solar
      - data_center
      - conservation
      - mineral_resource_screen
      - speculative_hold
      title: IntentCode
      type: string
...
    JobStatus:
      enum:
      - queued
      - running
      - succeeded
      - failed
      - cancelled
      - needs_review
      title: JobStatus
      type: string
...
    ReportReviewActionContract:
      properties:
        action:
          $ref: '#/components/schemas/ReportReviewStatus'
        from_status:
          $ref: '#/components/schemas/ReportReviewStatus'
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        reviewed_at:
          format: date-time
          title: Reviewed At
          type: string
        reviewer_id:
          title: Reviewer Id
          type: string
        to_status:
          $ref: '#/components/schemas/ReportReviewStatus'
      required:
      - action
      - from_status
      - to_status
      - reviewer_id
      title: ReportReviewActionContract
      type: object
    ReportReviewStatus:
      enum:
      - draft
      - needs_review
      - approved
      - rejected
      - superseded
      title: ReportReviewStatus
      type: string
    ReportRunContract:
      properties:
        area_id:
          format: uuid
          title: Area Id
          type: string
        artifact_metadata:
          additionalProperties: true
          title: Artifact Metadata
          type: object
        assumptions:
          items:
            type: string
          title: Assumptions
          type: array
        caveats:
          items:
            type: string
          title: Caveats
          type: array
        claims:
          items:
            $ref: '#/components/schemas/ClaimContract'
          title: Claims
          type: array
        evidence:
          items:
            $ref: '#/components/schemas/EvidenceContract'
          title: Evidence
          type: array
        finished_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished At
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        intent_code:
          $ref: '#/components/schemas/IntentCode'
        output_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Output Uri
        red_flags:
          items:
            $ref: '#/components/schemas/ClaimContract'
          title: Red Flags
          type: array
        report_run_id:
          format: uuid
          title: Report Run Id
          type: string
        requested_by:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Requested By
        review_actions:
          items:
            $ref: '#/components/schemas/ReportReviewActionContract'
          title: Review Actions
          type: array
        review_status:
          $ref: '#/components/schemas/ReportReviewStatus'
          default: needs_review
        reviewed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Reviewed At
        reviewed_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewed By
        source_manifest:
          additionalProperties: true
          title: Source Manifest
          type: object
        started_at:
          format: date-time
          title: Started At
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
          default: queued
        unknowns:
          items:
            $ref: '#/components/schemas/ClaimContract'
          title: Unknowns
          type: array
        verification_tasks:
          items:
            type: string
          title: Verification Tasks
          type: array
        workspace_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Workspace Id
      required:
      - area_id
      - intent_code
      title: ReportRunContract
      type: object
    ReportRunCreateRequest:
      properties:
        area_id:
          format: uuid
          title: Area Id
          type: string
        intent_code:
          $ref: '#/components/schemas/IntentCode'
      required:
      - area_id
      - intent_code
      title: ReportRunCreateRequest
      type: object
    SeverityBand:
      enum:
      - critical
      - high
      - medium
      - low
      - informational
      - unknown
      title: SeverityBand
      type: string
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Intent-aware land/locality due-diligence backend scaffold.
  title: land-diligence
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /report-runs:
    post:
      operationId: create_report_run_report_runs_post
      parameters:
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-Workspace-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRunCreateRequest'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/AsyncReportRunResponse'
                - $ref: '#/components/schemas/ReportRunContract'
                title: Response Create Report Run Report Runs Post
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Report Run
      tags:
      - report-runs
  /report-runs/{report_run_id}:
    get:
      operationId: get_report_run_report_runs__report_run_id__get
      parameters:
      - in: path
        name: report_run_id
        required: true
        schema:
          format: uuid
          title: Report Run Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: X-Workspace-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Workspace-Id
      - in: header
        name: X-User-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportRunContract'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Report Run
      tags:
      - report-runs
  /report-runs/{report_run_id}/retry:
    post:
      operationId: retry_report_run_report_runs__report_run_id__retry_post
      parameters:
      - in: path
        name: report_run_id
        required: true
        schema:
          format: uuid
          title: Report Run Id
          type: string
      - in: header
        name: X-Reviewer-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Reviewer-Id
      - in: header
        name: X-Reviewer-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Reviewer-Token
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncReportRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Retry Report Run
      tags:
      - report-runs
...